% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webfinger-package.R
\docType{package}
\name{webfinger}
\alias{webfinger}
\alias{webfinger-package}
\title{...}
\description{
A good description goes here otherwise CRAN checks fail.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hrbrmstr/webfinger}
  \item Report bugs at \url{https://github.com/webfinger/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
